src/requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distlib==0.3.7
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mysql-connector-python==8.2.0
packaging==23.2
platformdirs==4.1.0
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
virtualenv==20.25.0
Werkzeug==3.0.1
